+++ src/components/manage/Blocks/Block/QuantaBlockWrapper.jsx
@@ -3,14 +3,13 @@ import { Button } from 'semantic-ui-react';
 import includes from 'lodash/includes';
 import isBoolean from 'lodash/isBoolean';
 import { defineMessages, injectIntl } from 'react-intl';
-import { Icon, BlockChooser } from '@plone/volto/components';
+import { Icon } from '@plone/volto/components';
 import { blockHasValue } from '@plone/volto/helpers';
 import { Pluggable } from '@plone/volto/components/manage/Pluggable';
-import { doesNodeContainClick } from 'semantic-ui-react/dist/commonjs/lib';
+import MutateBlockButton from './MutateBlockButton';
 
 import config from '@plone/volto/registry';
 
-import addSVG from '@plone/volto/icons/circle-plus.svg';
 import dragSVG from '@plone/volto/icons/drag.svg';
 import trashSVG from '@plone/volto/icons/delete.svg';
 
@@ -20,56 +19,6 @@ const messages = defineMessages({
     defaultMessage: 'delete',
   },
 });
-
-const MutateBlockButton = ({ block, allowedBlocks, data, onMutateBlock }) => {
-  const { disableNewBlocks } = data;
-  const [addNewBlockOpened, setAddNewBlockOpened] = React.useState(false);
-
-  const blockChooserRef = React.useRef();
-
-  const handleClickOutside = React.useCallback((e) => {
-    if (
-      blockChooserRef.current &&
-      doesNodeContainClick(blockChooserRef.current, e)
-    )
-      return;
-    setAddNewBlockOpened(false);
-  }, []);
-
-  React.useEffect(() => {
-    document.addEventListener('mousedown', handleClickOutside, false);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside, false);
-    };
-  }, [handleClickOutside]);
-
-  return (
-    <>
-      {!disableNewBlocks && !blockHasValue(data) && (
-        <Button
-          icon
-          basic
-          title="Add block"
-          onClick={() => setAddNewBlockOpened(true)}
-          className="group-block-add-button"
-        >
-          <Icon name={addSVG} className="" size="19px" />
-        </Button>
-      )}
-      {addNewBlockOpened && (
-        <BlockChooser
-          onMutateBlock={(id, value) => {
-            setAddNewBlockOpened(false);
-            onMutateBlock(id, value);
-          }}
-          currentBlock={block}
-          allowedBlocks={allowedBlocks}
-          ref={blockChooserRef}
-        />
-      )}
-    </>
-  );
-};
 
 const EditBlockWrapper = (props) => {
   const hideHandler = (data) => {
@@ -103,7 +52,10 @@ const EditBlockWrapper = (props) => {
           >
             <Icon name={dragSVG} size="18px" />
           </Button>
-          <MutateBlockButton {...blockProps} />
+          <MutateBlockButton
+            {...blockProps}
+            className="quanta-block-add-button"
+          />
 
           {!required && (
             <Button